JNLP works on Mac's OS X.
Changed how initial icon is loaded.
Added overriding of prompt.wav to accomidate for temp directory.
I need help with the layout of session.txt to make it more position
independent.

The build.xml file is a simple ant script for signing JAR files. 
Handy with the JNLP stuff.  See http://MattPayne.org/mmi for the rest
of the JNLP stuff.