Put Brian's suggested BMP fix in; did not test with BMPs -- yet.
They use bmp images which don't display with the program right now. I
looked at the code and noticed that you don't use the 'img' field of
the
CategoryItem. You use the image file to get a path to the image and
read
it in again using an icon. I suspect that it will work better if you
use
the image that ImageIO already read into memory.